requirements.txt
streamlit
streamlit-server-state
streamlit-autorefresh
fastapi
uvicorn
python-socketio
streamlit-authenticator
pandas
pyyaml